refactor #28 : Funnel 컴포넌트 반환 형태 통일을 위한 정적 프로퍼티 추가 패턴으로 수정
