+++ src/components/Funnel/index.tsx
@@ -1,28 +1,18 @@
-import { Children, isValidElement, type ReactNode, type ReactElement } from 'react';
-import { type NonEmptyArray } from 'models';
+import { Children, isValidElement } from 'react';
+import { type NonEmptyArray, type FunnelProps, type StepProps } from 'models';
 
-export interface FunnelProps<Steps extends NonEmptyArray<string>> {
-  steps: Steps;
-  step: Steps[number];
-  children: Array<React.ReactElement<StepProps<Steps>>> | ReactElement<StepProps<Steps>>;
-}
-
-export interface StepProps<Steps extends NonEmptyArray<string>> {
-  name: Steps[number];
-  onEnter?: () => void;
-  children: ReactNode;
-}
-
-export const Funnel = <Steps extends NonEmptyArray<string>>({ step, steps, children }: FunnelProps<Steps>) => {
+export default function Funnel<Steps extends NonEmptyArray<string>>({ step, steps, children }: FunnelProps<Steps>) {
   const validChildren = Children.toArray(children)
     .filter(isValidElement)
     .filter((item) => steps.includes((item.props as Partial<StepProps<Steps>>).name ?? '')) as Array<
     React.ReactElement<StepProps<Steps>>
   >;
-  const targetStep = validChildren.find((child) => child.props.name === step);
-  return <>{targetStep}</>;
-};
 
-export function Step<Steps extends NonEmptyArray<string>>({ children }: StepProps<Steps>) {
-  return <>{children}</>;
+  const targetStep = validChildren.find((child) => child.props.name === step);
+
+  return <>{targetStep}</>;
 }
+
+Funnel.Step = function Step<Steps extends NonEmptyArray<string>>({ children }: StepProps<Steps>) {
+  return <>{children}</>;
+};
