+++ src/models/index.ts
@@ -1,2 +1,3 @@
 export * from './common';
 export * from './ladder';
+export * from './funnel';
